+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unpublished Changes
 
+- Fixed a bug where pipes that inherited transform() got flagged.
 - Fixed a bug where parts' templateUrls should be relative to the parts' library
   and not the part itself.
 - Added support for === operator.